Change dd_hash to 9x9 without wrap-around
A bit more accurate according to Neal Krawetz[0]

Doing both the row and the column hash in one pass over a 9x9 image
means that the pixels in the bottom row and the rightmost column don't
carry as much weight as the rest. If this turns out to be a problem in
practice, I'll change it. I find the current implementation pretty neat
though.

[0] http://www.hackerfactor.com/blog/?/archives/529-Kind-of-Like-That.html